+++ gensim/models/lsimodel.py
@@ -59,6 +59,8 @@ import scipy.sparse
 from scipy.sparse import sparsetools
 
 from gensim import interfaces, matutils, utils
+from gensim.models import basemodel
+
 from six import iterkeys
 from six.moves import xrange
 
@@ -221,7 +223,7 @@ class Projection(utils.SaveLoad):
 #endclass Projection
 
 
-class LsiModel(interfaces.TransformationABC):
+class LsiModel(interfaces.TransformationABC,basemodel.BaseTopicModel):
     """
     Objects of this class allow building and maintaining a model for Latent
     Semantic Indexing (also known as Latent Semantic Analysis).
@@ -489,16 +491,6 @@ class LsiModel(interfaces.TransformationABC):
         norm = numpy.sqrt(numpy.sum(numpy.dot(c, c)))
         most = matutils.argsort(numpy.abs(c), topn, reverse=True)
         return [(self.id2word[val], 1.0 * c[val] / norm) for val in most]
-
-    def print_topic(self, topicno, topn=10):
-        """
-        Return a single topic as a formatted string. See `show_topic()` for parameters.
-
-        >>> lsimodel.print_topic(10, topn=5)
-        '-0.340 * "category" + 0.298 * "$M$" + 0.183 * "algebra" + -0.174 * "functor" + -0.168 * "operator"'
-
-        """
-        return ' + '.join(['%.3f*"%s"' % (v, k) for k, v in self.show_topic(topicno, topn)])
 
     def show_topics(self, num_topics=-1, num_words=10, log=False, formatted=True):
         """
